switch ids payload to uint to better align with the base repo model

+++ dto/grocery_batch_delete.go
@@ -1,5 +1,5 @@
 package dto
 
 type GroceryBatchDeleteRequest struct {
-	IDs []int64 `json:"ids" validate:"required,min=1,max=300,dive,gt=0"`
+	IDs []uint `json:"ids" validate:"required,min=1,max=300,dive,gt=0"`
 }
